requirements.txt
tdaf-ost-service-common==0.1
PasteDeploy==1.5.0
Routes==1.13
SQLAlchemy>=0.7.8,<0.7.99
Tempita==0.5.1
WebOb==1.2.3
amqp==1.0.12
anyjson==0.3.3
argparse==1.2.1
d2to1==0.2.10
decorator==3.4.0
eventlet==0.13.0
greenlet==0.4.1
iso8601==0.1.4
kombu==2.5.12
lxml==3.2.1
oslo.config==1.1.1
pbr==0.5.19
prettytable==0.7.2
python-keystoneclient==0.3.1
qpid-python==0.20
repoze.lru==0.6
requests==1.2.3
setuptools-git==1.0
simplejson==3.3.0
six==1.3.0
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
pycrypto==2.6
python-heatclient==0.2.3

